+++ packages/expo-modules-core/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 🛠 Breaking changes
 
 - Convertible enums must inherit from `expo.modules.kotlin.types.Enumerable` on Android.
+- `AppContext.currentActivity` is not longer returning `AppCompatActivity`, but an instance of `android.app.Activity` class.
 
 ### 🎉 New features